Rewrite to extend GenerateMembersHandlerBase
